fresh dev branch + tests

+++ app/Http/Controllers/API/UserAuthenticationController.php
@@ -107,8 +107,8 @@ class UserAuthenticationController extends Controller
     public function logout(Request $request): JsonResponse
     {
         //$request->user()->tokens()->delete();
-        $user = $request->user();
-        $user->tokens()->where('id', $user->currentAccessToken()->id)->delete();
+
+        auth()->user()->tokens()->where('id', $request->user()->currentAccessToken()->id )->delete();
 
         return $this->respondWithSuccess(['message' => 'Logged out successfully']);
     }
